api/requirements.txt
requests==2.28.2
flask==3.0.0
Flask-Cors==4.0.0
DateTime==5.2
python-dotenv==1.0.0
coverage==7.3.2
flask-sqlalchemy==3.1.1
pymysql==1.1.0